+++ packages/headless/src/playwright.ts
@@ -1,2 +1,2 @@
-export { comboboxTesting } from "./composables/comboBox/TestDropdown.ct";
-export * from "./composables/listbox/TestListbox.ct";
+export * from "./composables/comboBox/createComboBox.ct";
+export * from "./composables/listbox/createListbox.ct";
